Fixed "select line <num>" command.
